fix(audit): 교차파일 결함 2차 수정 (카드이미지·토큰암호화·실시간인증·N+1)
버킷 경계로 1차에서 보류됐던 교차파일 결함 마무리:

- 카드 이미지 필드 배선(108/109): CardDeckItemDto imageStorageKey/imageUrl 응답 배선,
  merge-guest 이미지 보존, repository/service/request/테스트 일괄
- OAuth 토큰 평문 저장 차단(87): 공용 OAuthTokenSecretProtector(AES/GCM) 도입,
  Google/OneDrive oauth·browser 토큰 read/write 양쪽 봉투 암호화
- 카드방 참가자 가장 차단(166): 백엔드 HMAC participant 토큰 발급 +
  race-server가 SPRING_INTERNAL_TOKEN으로 (roomId,participantId) 검증, 위변조 차단
  (race-shared 계약·web/mobile 연결부 하위호환 배선)
- N+1/배치(63/73): space_templates 스냅샷 필드 1회 로드, member_fields reorder 배치+영향행 검증
- GoogleDrive size int→long 오버플로(99)

검증: backend compileJava/compileTestJava 0, typecheck 10/10, lint 10/10, verify:parity OK

+++ apps/backend/src/main/java/world/yeon/backend/googledrive_browser/service/GoogleDriveBrowserService.java
@@ -75,7 +75,7 @@ public class GoogleDriveBrowserService {
 				files.add(new GoogleDriveFileResponse(
 					file.path("id").asText(),
 					file.path("name").asText(),
-					parseInt(file.path("size").asText("0")),
+					parseLong(file.path("size").asText("0")),
 					file.path("modifiedTime").asText(),
 					file.path("mimeType").asText()
 				));
@@ -194,7 +194,7 @@ public class GoogleDriveBrowserService {
 		return value;
 	}
 	private String url(String value) { return URLEncoder.encode(value, StandardCharsets.UTF_8); }
-	private int parseInt(String value) { try { return Integer.parseInt(value); } catch (Exception e) { return 0; } }
+	private long parseLong(String value) { try { return Long.parseLong(value); } catch (Exception e) { return 0L; } }
 
 	public record DownloadedFile(byte[] bytes, String contentType) {}
 	private record RefreshedToken(String accessToken, String refreshToken, OffsetDateTime expiresAt) {}
